+++ plugins/postcss-double-position-gradients/src/has-supports-at-rule-ancestor.js
@@ -1,0 +1,19 @@
+import { includesGradientsFunction } from './is-gradient';
+
+export function hasSupportsAtRuleAncestor(node) {
+	let parent = node.parent;
+	while (parent) {
+		if (parent.type !== 'atrule') {
+			parent = parent.parent;
+			continue;
+		}
+
+		if (parent.name === 'supports' && includesGradientsFunction(parent.params)) {
+			return true;
+		}
+
+		parent = parent.parent;
+	}
+
+	return false;
+}